Separate common logic for LDAP group-provider
