+++ plugin/trino-password-authenticators/src/main/java/io/trino/plugin/password/ldap/LdapAuthenticator.java
@@ -14,7 +14,6 @@
 package io.trino.plugin.password.ldap;
 
 import com.google.common.annotations.VisibleForTesting;
-import com.google.common.base.CharMatcher;
 import com.google.common.cache.CacheBuilder;
 import com.google.common.cache.CacheLoader;
 import com.google.common.cache.LoadingCache;
@@ -32,11 +31,9 @@ import javax.naming.NamingException;
 import java.security.Principal;
 import java.util.List;
 import java.util.Optional;
-import java.util.Set;
 
 import static com.google.common.base.Preconditions.checkArgument;
 import static com.google.common.base.Throwables.throwIfInstanceOf;
-import static com.google.common.collect.Iterables.getOnlyElement;
 import static java.lang.String.format;
 import static java.util.Objects.requireNonNull;
 import static java.util.concurrent.TimeUnit.MILLISECONDS;
@@ -45,39 +42,27 @@ public class LdapAuthenticator
         implements PasswordAuthenticator
 {
     private static final Logger log = Logger.get(LdapAuthenticator.class);
-    private static final CharMatcher SPECIAL_CHARACTERS = CharMatcher.anyOf(",=+<>#;*()\"\\\u0000");
-    private static final CharMatcher WHITESPACE = CharMatcher.anyOf(" \r");
 
-    private final LdapAuthenticatorClient client;
+    private final LdapClient client;
+    private final LdapCommon ldapCommon;
 
     private final List<String> userBindSearchPatterns;
     private final Optional<String> groupAuthorizationSearchPattern;
     private final Optional<String> userBaseDistinguishedName;
-    private final Optional<String> bindDistinguishedName;
-    private final Optional<String> bindPassword;
 
     private final LoadingCache<Credential, Principal> authenticationCache;
 
     @Inject
-    public LdapAuthenticator(LdapAuthenticatorClient client, LdapConfig ldapConfig)
+    public LdapAuthenticator(LdapClient client, LdapConfig ldapConfig, LdapCommon ldapCommon)
     {
         this.client = requireNonNull(client, "client is null");
+        this.ldapCommon = ldapCommon;
 
         this.userBindSearchPatterns = ldapConfig.getUserBindSearchPatterns();
         this.groupAuthorizationSearchPattern = Optional.ofNullable(ldapConfig.getGroupAuthorizationSearchPattern());
         this.userBaseDistinguishedName = Optional.ofNullable(ldapConfig.getUserBaseDistinguishedName());
-        this.bindDistinguishedName = Optional.ofNullable(ldapConfig.getBindDistingushedName());
-        this.bindPassword = Optional.ofNullable(ldapConfig.getBindPassword());
+        Optional<String> bindDistinguishedName = Optional.ofNullable(ldapConfig.getBindDistingushedName());
 
-        checkArgument(
-                groupAuthorizationSearchPattern.isEmpty() || userBaseDistinguishedName.isPresent(),
-                "Base distinguished name (DN) for user must be provided");
-        checkArgument(
-                bindDistinguishedName.isPresent() == bindPassword.isPresent(),
-                "Both bind distinguished name and bind password must be provided together");
-        checkArgument(
-                bindDistinguishedName.isEmpty() || groupAuthorizationSearchPattern.isPresent(),
-                "Group authorization search pattern must be provided when bind distinguished name is used");
         checkArgument(
                 bindDistinguishedName.isPresent() || !userBindSearchPatterns.isEmpty(),
                 "Either user bind search pattern or bind distinguished name must be provided");
@@ -110,17 +95,17 @@ public class LdapAuthenticator
     private Principal authenticateWithUserBind(Credential credential)
     {
         String user = credential.getUser();
-        if (containsSpecialCharacters(user)) {
+        if (LdapCommon.containsSpecialCharacters(user)) {
             throw new AccessDeniedException("Username contains a special LDAP character");
         }
         Exception lastException = new RuntimeException();
         for (String userBindSearchPattern : userBindSearchPatterns) {
             try {
-                String userDistinguishedName = replaceUser(userBindSearchPattern, user);
+                String userDistinguishedName = LdapCommon.replaceUser(userBindSearchPattern, user);
                 if (groupAuthorizationSearchPattern.isPresent()) {
                     // user password is also validated as user DN and password is used for querying LDAP
                     String searchBase = userBaseDistinguishedName.orElseThrow();
-                    String groupSearch = replaceUser(groupAuthorizationSearchPattern.get(), user);
+                    String groupSearch = LdapCommon.replaceUser(groupAuthorizationSearchPattern.get(), user);
                     if (!client.isGroupMember(searchBase, groupSearch, userDistinguishedName, credential.getPassword())) {
                         String message = format("User [%s] not a member of an authorized group or not exist", user);
                         log.debug("%s", message);
@@ -147,11 +132,11 @@ public class LdapAuthenticator
     private Principal authenticateWithBindDistinguishedName(Credential credential)
     {
         String user = credential.getUser();
-        if (containsSpecialCharacters(user)) {
+        if (LdapCommon.containsSpecialCharacters(user)) {
             throw new AccessDeniedException("Username contains a special LDAP character");
         }
         try {
-            String userDistinguishedName = lookupUserDistinguishedName(user);
+            String userDistinguishedName = ldapCommon.lookupUserDistinguishedName(user);
             client.validatePassword(userDistinguishedName, credential.getPassword());
             log.debug("Authentication successful for user [%s]", user);
         }
@@ -160,45 +145,5 @@ public class LdapAuthenticator
             throw new RuntimeException("Authentication error");
         }
         return new BasicPrincipal(credential.getUser());
-    }
-
-    /**
-     * Returns {@code true} when parameter contains a character that has a special meaning in
-     * LDAP search or bind name (DN).
-     * <p>
-     * Based on <a href="https://www.owasp.org/index.php/Preventing_LDAP_Injection_in_Java">Preventing_LDAP_Injection_in_Java</a> and
-     * {@link javax.naming.ldap.Rdn#escapeValue(Object) escapeValue} method.
-     */
-    @VisibleForTesting
-    static boolean containsSpecialCharacters(String user)
-    {
-        if (WHITESPACE.indexIn(user) == 0 || WHITESPACE.lastIndexIn(user) == user.length() - 1) {
-            return true;
-        }
-        return SPECIAL_CHARACTERS.matchesAnyOf(user);
-    }
-
-    private String lookupUserDistinguishedName(String user)
-            throws NamingException
-    {
-        String searchBase = userBaseDistinguishedName.orElseThrow();
-        String searchFilter = replaceUser(groupAuthorizationSearchPattern.orElseThrow(), user);
-        Set<String> userDistinguishedNames = client.lookupUserDistinguishedNames(searchBase, searchFilter, bindDistinguishedName.orElseThrow(), bindPassword.orElseThrow());
-        if (userDistinguishedNames.isEmpty()) {
-            String message = format("User [%s] not a member of an authorized group", user);
-            log.debug("%s", message);
-            throw new AccessDeniedException(message);
-        }
-        if (userDistinguishedNames.size() > 1) {
-            String message = format("Multiple group membership results for user [%s]: %s", user, userDistinguishedNames);
-            log.debug("%s", message);
-            throw new AccessDeniedException(message);
-        }
-        return getOnlyElement(userDistinguishedNames);
-    }
-
-    private static String replaceUser(String pattern, String user)
-    {
-        return pattern.replace("${USER}", user);
     }
 }
