+++ plugin/trino-password-authenticators/src/main/java/io/trino/plugin/password/ldap/LdapAuthenticatorFactory.java
@@ -39,7 +39,7 @@ public class LdapAuthenticatorFactory
                 binder -> {
                     configBinder(binder).bindConfig(LdapConfig.class);
                     binder.bind(LdapAuthenticator.class).in(Scopes.SINGLETON);
-                    binder.bind(LdapAuthenticatorClient.class).to(JdkLdapAuthenticatorClient.class).in(Scopes.SINGLETON);
+                    binder.bind(LdapClient.class).to(JdkLdapClient.class).in(Scopes.SINGLETON);
                 });
 
         Injector injector = app
