+++ plugin/trino-password-authenticators/src/main/java/io/trino/plugin/password/ldap/LdapClient.java
@@ -17,7 +17,7 @@ import javax.naming.NamingException;
 
 import java.util.Set;
 
-public interface LdapAuthenticatorClient
+public interface LdapClient
 {
     void validatePassword(String userDistinguishedName, String password)
             throws NamingException;
