+++ plugin/trino-password-authenticators/src/test/java/io/trino/plugin/password/ldap/TestLdapAuthenticator.java
@@ -13,38 +13,25 @@
  */
 package io.trino.plugin.password.ldap;
 
-import com.google.common.collect.HashMultimap;
-import com.google.common.collect.ImmutableSet;
-import io.trino.plugin.password.Credential;
 import io.trino.spi.security.AccessDeniedException;
 import io.trino.spi.security.BasicPrincipal;
 import org.testng.annotations.Test;
 
-import javax.naming.NamingException;
-
-import java.util.HashSet;
-import java.util.Optional;
-import java.util.Set;
-
-import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static org.testng.Assert.assertEquals;
 
 public class TestLdapAuthenticator
 {
-    private static final String BASE_DN = "base-dn";
-    private static final String PATTERN_PREFIX = "pattern::";
-
     @Test
     public void testSingleBindPattern()
     {
-        TestLdapAuthenticatorClient client = new TestLdapAuthenticatorClient();
+        TestLdapClient client = new TestLdapClient();
         client.addCredentials("alice@example.com", "alice-pass");
+        LdapConfig ldapConfig = new LdapConfig()
+                .setUserBindSearchPatterns("${USER}@example.com");
+        LdapCommon ldapCommon = new LdapCommon(client, ldapConfig);
 
-        LdapAuthenticator ldapAuthenticator = new LdapAuthenticator(
-                client,
-                new LdapConfig()
-                        .setUserBindSearchPatterns("${USER}@example.com"));
+        LdapAuthenticator ldapAuthenticator = new LdapAuthenticator(client, ldapConfig, ldapCommon);
 
         assertThatThrownBy(() -> ldapAuthenticator.createAuthenticatedPrincipal("alice", "invalid"))
                 .isInstanceOf(RuntimeException.class);
@@ -56,12 +43,12 @@ public class TestLdapAuthenticator
     @Test
     public void testMultipleBindPattern()
     {
-        TestLdapAuthenticatorClient client = new TestLdapAuthenticatorClient();
+        TestLdapClient client = new TestLdapClient();
+        LdapConfig ldapConfig = new LdapConfig()
+                .setUserBindSearchPatterns("${USER}@example.com:${USER}@alt.example.com");
+        LdapCommon ldapCommon = new LdapCommon(client, ldapConfig);
 
-        LdapAuthenticator ldapAuthenticator = new LdapAuthenticator(
-                client,
-                new LdapConfig()
-                        .setUserBindSearchPatterns("${USER}@example.com:${USER}@alt.example.com"));
+        LdapAuthenticator ldapAuthenticator = new LdapAuthenticator(client, ldapConfig, ldapCommon);
 
         client.addCredentials("alice@example.com", "alice-pass");
         assertEquals(ldapAuthenticator.createAuthenticatedPrincipal("alice", "alice-pass"), new BasicPrincipal("alice"));
@@ -81,15 +68,15 @@ public class TestLdapAuthenticator
     @Test
     public void testGroupMembership()
     {
-        TestLdapAuthenticatorClient client = new TestLdapAuthenticatorClient();
+        TestLdapClient client = new TestLdapClient();
         client.addCredentials("alice@example.com", "alice-pass");
+        LdapConfig ldapConfig = new LdapConfig()
+                .setUserBindSearchPatterns("${USER}@example.com")
+                .setUserBaseDistinguishedName(TestLdapClient.BASE_DN)
+                .setGroupAuthorizationSearchPattern(TestLdapClient.PATTERN_PREFIX + "${USER}");
+        LdapCommon ldapCommon = new LdapCommon(client, ldapConfig);
 
-        LdapAuthenticator ldapAuthenticator = new LdapAuthenticator(
-                client,
-                new LdapConfig()
-                        .setUserBindSearchPatterns("${USER}@example.com")
-                        .setUserBaseDistinguishedName(BASE_DN)
-                        .setGroupAuthorizationSearchPattern(PATTERN_PREFIX + "${USER}"));
+        LdapAuthenticator ldapAuthenticator = new LdapAuthenticator(client, ldapConfig, ldapCommon);
 
         assertThatThrownBy(() -> ldapAuthenticator.createAuthenticatedPrincipal("alice", "invalid"))
                 .isInstanceOf(RuntimeException.class);
@@ -104,16 +91,16 @@ public class TestLdapAuthenticator
     @Test
     public void testDistinguishedNameLookup()
     {
-        TestLdapAuthenticatorClient client = new TestLdapAuthenticatorClient();
+        TestLdapClient client = new TestLdapClient();
         client.addCredentials("alice@example.com", "alice-pass");
+        LdapConfig ldapConfig = new LdapConfig()
+                .setUserBaseDistinguishedName(TestLdapClient.BASE_DN)
+                .setGroupAuthorizationSearchPattern(TestLdapClient.PATTERN_PREFIX + "${USER}")
+                .setBindDistingushedName("server")
+                .setBindPassword("server-pass");
+        LdapCommon ldapCommon = new LdapCommon(client, ldapConfig);
 
-        LdapAuthenticator ldapAuthenticator = new LdapAuthenticator(
-                client,
-                new LdapConfig()
-                        .setUserBaseDistinguishedName(BASE_DN)
-                        .setGroupAuthorizationSearchPattern(PATTERN_PREFIX + "${USER}")
-                        .setBindDistingushedName("server")
-                        .setBindPassword("server-pass"));
+        LdapAuthenticator ldapAuthenticator = new LdapAuthenticator(client, ldapConfig, ldapCommon);
 
         client.addCredentials("alice", "alice-pass");
         client.addCredentials("alice@example.com", "alice-pass");
@@ -136,100 +123,5 @@ public class TestLdapAuthenticator
         client.addDistinguishedNameForUser("alice", "another-mapping");
         assertThatThrownBy(() -> ldapAuthenticator.createAuthenticatedPrincipal("alice", "alice-pass"))
                 .isInstanceOf(AccessDeniedException.class);
-    }
-
-    @Test
-    public void testContainsSpecialCharacters()
-    {
-        assertThat(LdapAuthenticator.containsSpecialCharacters("The quick brown fox jumped over the lazy dogs"))
-                .as("English pangram")
-                .isEqualTo(false);
-        assertThat(LdapAuthenticator.containsSpecialCharacters("Pchnąć w tę łódź jeża lub ośm skrzyń fig"))
-                .as("Perfect polish pangram")
-                .isEqualTo(false);
-        assertThat(LdapAuthenticator.containsSpecialCharacters("いろはにほへと ちりぬるを わかよたれそ つねならむ うゐのおくやま けふこえて あさきゆめみし ゑひもせす（ん）"))
-                .as("Japanese hiragana pangram - Iroha")
-                .isEqualTo(false);
-        assertThat(LdapAuthenticator.containsSpecialCharacters("*"))
-                .as("LDAP wildcard")
-                .isEqualTo(true);
-        assertThat(LdapAuthenticator.containsSpecialCharacters("   John Doe"))
-                .as("Beginning with whitespace")
-                .isEqualTo(true);
-        assertThat(LdapAuthenticator.containsSpecialCharacters("John Doe  \r"))
-                .as("Ending with whitespace")
-                .isEqualTo(true);
-        assertThat(LdapAuthenticator.containsSpecialCharacters("Hi (This) = is * a \\ test # ç à ô"))
-                .as("Multiple special characters")
-                .isEqualTo(true);
-        assertThat(LdapAuthenticator.containsSpecialCharacters("John\u0000Doe"))
-                .as("NULL character")
-                .isEqualTo(true);
-        assertThat(LdapAuthenticator.containsSpecialCharacters("John Doe <john.doe@company.com>"))
-                .as("Angle brackets")
-                .isEqualTo(true);
-    }
-
-    private static class TestLdapAuthenticatorClient
-            implements LdapAuthenticatorClient
-    {
-        private final Set<Credential> credentials = new HashSet<>();
-        private final Set<String> groupMembers = new HashSet<>();
-        private final HashMultimap<String, String> userDNs = HashMultimap.create();
-
-        public void addCredentials(String userDistinguishedName, String password)
-        {
-            credentials.add(new Credential(userDistinguishedName, password));
-        }
-
-        public void addGroupMember(String userName)
-        {
-            groupMembers.add(userName);
-        }
-
-        public void addDistinguishedNameForUser(String userName, String distinguishedName)
-        {
-            userDNs.put(userName, distinguishedName);
-        }
-
-        @Override
-        public void validatePassword(String userDistinguishedName, String password)
-                throws NamingException
-        {
-            if (!credentials.contains(new Credential(userDistinguishedName, password))) {
-                throw new NamingException();
-            }
-        }
-
-        @Override
-        public boolean isGroupMember(String searchBase, String groupSearch, String contextUserDistinguishedName, String contextPassword)
-                throws NamingException
-        {
-            validatePassword(contextUserDistinguishedName, contextPassword);
-            return getSearchUser(searchBase, groupSearch)
-                    .map(groupMembers::contains)
-                    .orElse(false);
-        }
-
-        @Override
-        public Set<String> lookupUserDistinguishedNames(String searchBase, String searchFilter, String contextUserDistinguishedName, String contextPassword)
-                throws NamingException
-        {
-            validatePassword(contextUserDistinguishedName, contextPassword);
-            return getSearchUser(searchBase, searchFilter)
-                    .map(userDNs::get)
-                    .orElse(ImmutableSet.of());
-        }
-
-        private static Optional<String> getSearchUser(String searchBase, String groupSearch)
-        {
-            if (!searchBase.equals(BASE_DN)) {
-                return Optional.empty();
-            }
-            if (!groupSearch.startsWith(PATTERN_PREFIX)) {
-                return Optional.empty();
-            }
-            return Optional.of(groupSearch.substring(PATTERN_PREFIX.length()));
-        }
     }
 }
